Move mask-fetching func to regression test file where used
Only place it is used.
