+++ pm_icecon/tests/regression/test_bt.py
@@ -6,18 +6,70 @@ import numpy as np
 import xarray as xr
 from numpy.testing import assert_almost_equal
 from numpy.typing import NDArray
+import numpy.typing as npt
 from pm_tb_data.fetch.au_si import AU_SI_RESOLUTIONS
 from pm_tb_data._types import Hemisphere
 
 import pm_icecon.bt.compute_bt_ic as bt
 from pm_icecon.bt.api import amsr2_goddard_bootstrap
 from pm_icecon.bt.compute_bt_ic import xfer_class_tbs
-from pm_icecon.bt.masks import get_ps_invalid_ice_mask
+from pm_icecon.bt.masks import (
+    get_ps_invalid_ice_mask,
+    get_pss_12_validice_land_coast_array,
+)
 from pm_icecon.bt.params.class_sats import SSMIS_NORTH_PARAMS
 from pm_icecon.config.models.bt import BootstrapParams
-from pm_icecon.constants import CDR_TESTDATA_DIR
 from pm_icecon.interpolation import spatial_interp_tbs
-from pm_icecon.masks import get_ps_land_mask, get_ps_pole_hole_mask
+from pm_icecon.util import get_ps25_grid_shape
+from pm_icecon.constants import (
+    BT_GODDARD_ANCILLARY_DIR,
+    CDR_TESTDATA_DIR,
+)
+
+
+def _get_ps_land_mask(
+    *,
+    hemisphere: Hemisphere,
+    resolution: AU_SI_RESOLUTIONS,
+) -> npt.NDArray[np.bool_]:
+    """Get the polar stereo 25km land mask."""
+    # Ocean has a value of 0, land a value of 1, and coast a value of 2.
+    if resolution == "25":
+        shape = get_ps25_grid_shape(hemisphere=hemisphere)
+        _land_coast_array = np.fromfile(
+            (
+                BT_GODDARD_ANCILLARY_DIR
+                / (
+                    f"{hemisphere}_land_25"
+                    # NOTE: According to scotts, the 'r' in the southern hemisphere
+                    # filename probably stands for “revised“.
+                    f"{'r' if hemisphere == 'south' else ''}"
+                )
+            ).resolve(),
+            dtype=np.int16,
+        ).reshape(shape)
+
+        land_mask = _land_coast_array != 0
+
+        # TODO: land mask currently includes land and coast. Does this make sense? Are
+        # we ever going to need to coast values? Maybe rename to `LAND_COAST_MASK`?
+    elif resolution == "12":
+        if hemisphere == "south":
+            # Any date is OK. The land mask is the same for all of the pss 12
+            # validice/land masks
+            _land_coast_array = get_pss_12_validice_land_coast_array(
+                date=dt.date.today()
+            )
+            land_mask = np.logical_or(_land_coast_array == 0, _land_coast_array == 32)
+        else:
+            _land_coast_array = np.fromfile(
+                CDR_TESTDATA_DIR / "btequiv_psn12.5/bt_landequiv_psn12.5km.dat",
+                dtype=np.int16,
+            ).reshape(896, 608)
+
+            land_mask = _land_coast_array != 0
+
+    return land_mask
 
 
 def test_bt_amsr2_regression():
@@ -32,9 +84,8 @@ def test_bt_amsr2_regression():
     """
     resolution: Final = "25"
     hemisphere: Final = "north"
-    pole_mask = get_ps_pole_hole_mask(resolution=resolution)
 
-    land_mask = get_ps_land_mask(hemisphere=hemisphere, resolution=resolution)
+    land_mask = _get_ps_land_mask(hemisphere=hemisphere, resolution=resolution)
     for date in (dt.date(2020, 1, 1), dt.date(2022, 5, 4)):
         invalid_ice_mask = get_ps_invalid_ice_mask(
             hemisphere=hemisphere,
@@ -47,7 +98,6 @@ def test_bt_amsr2_regression():
             resolution=resolution,
             land_mask=land_mask,
             invalid_ice_mask=invalid_ice_mask,
-            pole_mask=pole_mask,
         )
 
         filename = f"NH_{date:%Y%m%d}_py_NRT_amsr2.nc"
@@ -83,7 +133,7 @@ def _original_f18_example() -> xr.Dataset:
     date = dt.date(2018, 2, 17)
     hemisphere: Hemisphere = "north"
 
-    land_mask = get_ps_land_mask(hemisphere=hemisphere, resolution=resolution)
+    land_mask = _get_ps_land_mask(hemisphere=hemisphere, resolution=resolution)
     invalid_ice_mask = get_ps_invalid_ice_mask(
         hemisphere=hemisphere,
         date=date,
